requirements.txt
Jinja2==2.8
MarkupSafe==0.23
PasteDeploy==1.5.2
PdbSublimeTextSupport==0.2
psycopg2==2.6.1
py==1.4.30
pyramid==1.5.7
pyramid-jinja2==2.5
pyramid-tm==0.12
pytest==2.8.2
repoze.lru==0.6
SQLAlchemy==1.0.9
transaction==1.4.4
translationstring==1.3
venusian==1.0
waitress==0.8.10
WebOb==1.4.1
zope.deprecation==4.1.2
zope.interface==4.1.3
zope.sqlalchemy==0.7.6
